DFA-2123 - refactoring and review comments
