+++ express/.env.example
@@ -1,7 +1,7 @@
 USERPOOL_ID= # Can be found at https://eu-west-2.console.aws.amazon.com/cognito/v2/idp/user-pools?region=eu-west-2 in the second column
 CLIENT_ID= # Can be found at https://eu-west-2.console.aws.amazon.com/cognito/v2/idp/user-pools/<above user pool ID>/app-integration?region=eu-west-2 at the bottom of the page in the second column
 API_BASE_URL= # API ID can be found at https://eu-west-2.console.aws.amazon.com/apigateway/main/apis?region=eu-west-2 in the third column.  Substitute the value into this: https://<API ID goes here>.execute-api.eu-west-2.amazonaws.com
-SESSION_STORAGE_ENABLED= # true or false.Default in memory implementation will be used if it is false.
+SESSION_STORAGE= # DynamoDB any other value will use in memory session storage.
 SESSIONS_TABLE= # Sessions table name can be found at https://eu-west-2.console.aws.amazon.com/dynamodbv2/home?region=eu-west-2#tables. It is one of the the table in the list with this pattern <DeployedStackName>-Sessions-XXXXXXX
-SESSION_LOCAL= # true/false - true if you need to use session storage on your local machine. Please follow instructions in README.md for installing and running docker image locally.
+DYNAMO_DB_ENDPOINT= # http://localhost:8000, if you need to overwrite to local or any other dynamoDB endpoint.
 PORT=3000
